<refactor>[Products]: components ts with separete settings
