+++ src/app/products/products.component.ts
@@ -11,7 +11,8 @@ import { ProductsService } from '../services/products/products.service';
 
 import { Product } from '../interfaces/product.interface';
 
-import { AddOrEditProductComponent } from './add-or-edit-product/add-or-edit-product.component';
+import { CreateProductComponent } from './create-product/create-product.component';
+import { UpdateProductComponent } from './update-product/update-product.component';
 import { DeleteProductComponent } from './delete-product/delete-product.component';
 
 import { MOCK_PRODUCTS } from './products.mock';
@@ -148,11 +149,11 @@ export class ProductsComponent implements OnInit, AfterViewInit, OnDestroy {
     if(product !== null) {
       products.push(product);
 
-      this.dialog.open<AddOrEditProductComponent>(AddOrEditProductComponent, {
+      this.dialog.open<UpdateProductComponent>(UpdateProductComponent, {
         data: products
       });
     } else {
-      this.dialog.open<AddOrEditProductComponent>(AddOrEditProductComponent);
+      this.dialog.open<CreateProductComponent>(CreateProductComponent);
     };
   }
